+++ R/row_sums.R
@@ -7,7 +7,7 @@
 #' @param ... The columns to sum, take the mean of, etc. \emph{Required}.
 #' @param .fun The function to apply.
 #' @param na_rm Whether to remove \code{NA} values or not. Default is \code{FALSE}.
-#' @param varname The column name of the sums means etc.
+#' @param varname The column name of the sums means etc. as a character string.
 #'
 #' @details Simple wrappers for applying rowwise operations only for selected
 #'   columns within the tidyverse approach to data processing.  The
@@ -28,6 +28,9 @@
 #' d  %>%
 #'   row_means(matches('x|z'))
 #'
+#' d  %>%
+#'   row_max(matches('x|y'))
+#'
 #' row_apply(
 #'   d ,
 #'   everything(),
@@ -37,6 +40,8 @@
 #'
 #' @export
 row_sums <- function(data, ..., na_rm = FALSE, varname = 'sum') {
+  # note: dplyr 1.0 included rowwise operations, but it wasn't obvious what
+  # advantage there would be for these functions except for min and max
   dplyr::mutate(data, !!varname := rowSums(select(data, ...), na.rm = na_rm))
 }
 
@@ -44,6 +49,28 @@ row_sums <- function(data, ..., na_rm = FALSE, varname = 'sum') {
 #' @rdname row_sums
 row_means <- function(data, ..., na_rm = FALSE, varname = 'mean') {
   dplyr::mutate(data, !!varname := rowMeans(select(data, ...), na.rm = na_rm))
+}
+
+#' @export
+#' @rdname row_sums
+row_min <- function(data, ..., na_rm = FALSE, varname = 'min') {
+  dplyr::select(data, ...) %>%
+    dplyr::mutate(rn = 1:nrow(.)) %>%
+    dplyr::rowwise(rn) %>%
+    dplyr::mutate(!!varname := min(dplyr::c_across(...), na.rm = na_rm)) %>%
+    dplyr::ungroup() %>%
+    dplyr::select(-rn)
+}
+
+#' @export
+#' @rdname row_sums
+row_max <- function(data, ..., na_rm = FALSE, varname = 'max') {
+  dplyr::select(data, ...) %>%
+    dplyr::mutate(rn = 1:nrow(.)) %>%
+    dplyr::rowwise(rn) %>%
+    dplyr::mutate(!!varname := max(dplyr::c_across(...), na.rm = na_rm)) %>%
+    dplyr::ungroup() %>%
+    dplyr::select(-rn)
 }
 
 #' @export