+++ R/utils.R
@@ -1,9 +1,1 @@
-#' @importFrom magrittr %>%
-#' @export
-magrittr::`%>%`
-
-
-
-#' @importFrom dplyr vars
-#' @export
-dplyr::vars
+#' @importFrom dplyr vars `%>%` tibble