+++ tests/testthat/test_cat_by.R
@@ -1,7 +1,7 @@
 context('test cat_by')
 
 set.seed(1)
-df1 <- tibble(
+df1 <- dplyr::tibble(
   g1 = factor(sample(1:2, 50, replace = TRUE), labels=c('a','b')),
   g2 = sample(1:4, 50, replace = TRUE),
   a = rnorm(50),