+++ tests/testthat/test_describe_all.R
@@ -2,7 +2,7 @@ context('test describe_all')
 
 set.seed(1234)
 
-df1 <- tibble(
+df1 <- dplyr::tibble(
   g1 = factor(sample(1:2, 50, replace = TRUE), labels=c('a','b')),
   g2 = sample(1:4, 50, replace = TRUE),
   g3 = ordered(sample(letters[1:3], 50, replace = TRUE)),