+++ tests/testthat/test_pre_process.R
@@ -2,7 +2,8 @@
 context('test pre_process')
 
 set.seed(1234)
-df1 <- tibble(
+
+df1 <- dplyr::tibble(
   g1 = factor(sample(1:2, 50, replace = TRUE), labels=c('a','b')),
   g2 = sample(1:4, 50, replace = TRUE),
   a = rnorm(50),